pyproject.toml
[tool.poetry]
name = "squawker"
version = "0.1.0"
description = ""
authors = ["3VAbdAVE <david@maskme.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
"RPi.GPIO" = "^0.7.1"
adafruit-circuitpython-motorkit = "^1.6.11"
argparse = "^1.4.0"
Wave = "^0.0.2"
numpy = "^1.23.3"
PyAudio = "^0.2.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
